requirements.txt
click
pyperclip
Pillow
loguru
requests